Made significant progress with the project. Fixed camera and movements issues. Collisions are also handled but might need some minor rework to be perfect. Handled the transition of 3d to 2d window as well.
